\name{CELLector.HCCancerDrivers}
\alias{CELLector.HCCancerDrivers}
\docType{data}
\title{
High Confidence Cancer Driver genes
}
\description{
A list of high confidence cancer driver genes from [1]
}
\usage{data("CELLector.HCCancerDrivers")}
\format{
 A vector of strings with one entry per cancer gene.
}
\references{
[1] Iorio, F. et al. A Landscape of Pharmacogenomic Interactions in Cancer. Cell 166, 740–754 (2016).
}
\examples{
data(CELLector.HCCancerDrivers)
## maybe str(CELLector.HCCancerDrivers) ; plot(CELLector.HCCancerDrivers) ...
}
\keyword{datasets}
